доделал апи

+++ src/utils/api.js
@@ -26,25 +26,25 @@ class Api {
         .then(this._handleResponse);
     }
 
-    editProfileInfo(newInfo) {
+    editProfileInfo(name, about) {
         return fetch(this.url + '/users/me', {
             method: 'PATCH',
             headers: this._headers,
             body: JSON.stringify({
-                name: newInfo.nickName,
-                about: newInfo.info
+                name: name,
+                about: about
             })
         })
         .then(this._handleResponse);
     }
 
-    createCard(data) {
+    createCard(name, link) {
         return fetch(this.url + '/cards', {
             method: 'POST',
             headers: this._headers,
             body: JSON.stringify({
-                name: data.name,
-                link: data.link
+                name: name,
+                link: link
             })
         })
         .then(this._handleResponse);
@@ -58,28 +58,29 @@ class Api {
         .then(this._handleResponse);
     }
 
-    putLike(id) {
-        return fetch(this.url + /cards/ + `${id}` + '/likes', {
-            method: 'PUT',
-            headers: this._headers
-        })
-        .then(this._handleResponse);
-    }
-
-    removeLike(id) {
-        return fetch(this.url + /cards/ + `${id}` + '/likes', {
-            method: 'DELETE',
-            headers: this._headers
-        })
-        .then(this._handleResponse);
+    changeLikeCardStatus(id, statusLike) {
+        if (statusLike === false) {
+            return fetch(this.url + /cards/ + `${id}` + '/likes', {
+                method: 'DELETE',
+                headers: this._headers
+            })
+            .then(this._handleResponse);
+        } else {
+            return fetch(this.url + /cards/ + `${id}` + '/likes', {
+                method: 'PUT',
+                headers: this._headers
+            })
+            .then(this._handleResponse);
+        }
     }
 
     updateAvatar(newAvatar) {
+        console.log(newAvatar);
         return fetch(this.url + '/users/me/avatar ', {
             method: 'PATCH',
             headers: this._headers,
             body: JSON.stringify({
-                avatar: newAvatar.avatar
+                avatar: newAvatar
             })
         })
         .then(this._handleResponse);
